implementation of sha256(sha256($pass))
Not-yet-working, mechanical conversion of mode 4500 to mode 1470, since
mode 1400 (sha256) is basically a find-n-replace on mode 100 (sha1), at
least according to this one-liner.

    sed -e 's/sha1/sha256/g' -e 's/m00100/m01400/g' < m00100_a3.cl | diff -u - m01400_a3.cl

It could just be broken on my system, since I'm running on an Intel GPU
with --force. But `./hashcat -b -m 100 --force` works, so I'm thinking
that this CL file is somehow incorrect.