Fix potfile line splitting for LM
Lines in a potfile with LM hashes have the format <hash>:<pw>. Splitting for --show is done by finding the last occurrence of the separator : in a line.
This breaks found password with a : in it as they will be marked as [notfound].

Fix this by looking for the FIRST occurrence instead of the LAST one (strchr vs strrchr).